requirements.txt
prompt-toolkit
attrs
